requirements.txt
asgiref==3.2.10
Django==3.1.1
django-extensions==3.0.8
django-filter==2.3.0
djangorestframework==3.11.1
Faker==4.1.3
flake8==3.8.3
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
yapf==0.30.0
